feat: remove project description min length + refactor project forms + style EditProjectForm
