+++ dev-client/src/screens/CreateProjectScreen/components/ProjectForm.tsx
@@ -15,19 +15,12 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {useCallback} from 'react';
 import {useTranslation} from 'react-i18next';
 
-import {Formik, FormikHelpers, useFormikContext} from 'formik';
+import {useFormikContext} from 'formik';
 import {TFunction} from 'i18next';
 import * as yup from 'yup';
 
-import {
-  ProjectMembershipProjectRoleChoices,
-  ProjectUpdateMutationInput,
-} from 'terraso-client-shared/graphqlSchema/graphql';
-
-import {ContainedButton} from 'terraso-mobile-client/components/buttons/ContainedButton';
 import {HelpContentSpacer} from 'terraso-mobile-client/components/content/HelpContentSpacer';
 import {DataPrivacyInfoButton} from 'terraso-mobile-client/components/content/info/privacy/DataPrivacyInfoButton';
 import {FormTextField} from 'terraso-mobile-client/components/form/FormTextField';
@@ -35,12 +28,10 @@ import {
   Column,
   Heading,
   Row,
-  View,
 } from 'terraso-mobile-client/components/NativeBaseAdapters';
 import {RadioBlock} from 'terraso-mobile-client/components/RadioBlock';
 import {
   PROJECT_DESCRIPTION_MAX_LENGTH,
-  PROJECT_DESCRIPTION_MIN_LENGTH,
   PROJECT_NAME_MAX_LENGTH,
   PROJECT_NAME_MIN_LENGTH,
 } from 'terraso-mobile-client/constants';
@@ -62,30 +53,17 @@ export const projectValidationFields = (t: TFunction) => ({
       }),
     )
     .required(t('general.required')),
-  description: yup
-    .string()
-    .min(
-      PROJECT_DESCRIPTION_MIN_LENGTH,
-      t('projects.form.description_min_length_error', {
-        max: PROJECT_DESCRIPTION_MIN_LENGTH,
-      }),
-    )
-    .max(
-      PROJECT_DESCRIPTION_MAX_LENGTH,
-      t('projects.form.description_max_length_error', {
-        max: PROJECT_DESCRIPTION_MAX_LENGTH,
-      }),
-    ),
+  description: yup.string().max(
+    PROJECT_DESCRIPTION_MAX_LENGTH,
+    t('projects.form.description_max_length_error', {
+      max: PROJECT_DESCRIPTION_MAX_LENGTH,
+    }),
+  ),
   privacy: yup.string().oneOf(['PRIVATE', 'PUBLIC']).required(),
 });
 
 export const projectValidationSchema = (t: TFunction) =>
   yup.object().shape(projectValidationFields(t));
-
-export const editProjectValidationSchema = (t: TFunction) => {
-  const fullSchema = projectValidationSchema(t);
-  return fullSchema.pick(['name', 'description']);
-};
 
 export type ProjectFormValues = {
   name: string;
@@ -93,77 +71,12 @@ export type ProjectFormValues = {
   privacy: 'PUBLIC' | 'PRIVATE';
 };
 
-type FormValues = Omit<ProjectUpdateMutationInput, 'id'>;
-
-type FormProps = FormValues & {
-  onSubmit: (values: FormValues) => Promise<void>;
-  userRole: ProjectMembershipProjectRoleChoices | null;
-};
-
-export const EditProjectForm = ({
-  onSubmit,
-  name,
-  description,
-  userRole,
-}: Omit<FormProps, 'privacy'>) => {
-  const {t} = useTranslation();
-
-  /* We reset the form after a successful submit to clear the dirty flag */
-  const onSubmitAndReset = useCallback(
-    (values: FormValues, helpers: FormikHelpers<FormValues>) => {
-      onSubmit(values).then(() => {
-        helpers.resetForm({values});
-      });
-    },
-    [onSubmit],
-  );
-
-  return (
-    <Formik<FormValues>
-      validationSchema={editProjectValidationSchema(t)}
-      initialValues={{name, description}}
-      validateOnMount={true}
-      onSubmit={onSubmitAndReset}>
-      {({handleSubmit, isValid, isSubmitting, dirty}) => (
-        <>
-          <FormTextField<FormValues>
-            name="name"
-            maxLength={PROJECT_NAME_MAX_LENGTH}
-            placeholder={t('projects.create.name_label')}
-            label={t('projects.create.name_label')}
-            showCounter
-          />
-          <FormTextField<FormValues>
-            name="description"
-            maxLength={PROJECT_DESCRIPTION_MAX_LENGTH}
-            placeholder={t('projects.create.description_label')}
-            label={t('projects.create.description_label')}
-            numberOfLines={3}
-            multiline
-            showCounter
-          />
-          {userRole === 'MANAGER' && (
-            <View flexDirection="row" justifyContent="flex-end">
-              <ContainedButton
-                size="lg"
-                onPress={handleSubmit}
-                disabled={!dirty || isSubmitting || !isValid}
-                label={t('general.save')}
-              />
-            </View>
-          )}
-        </>
-      )}
-    </Formik>
-  );
-};
-
 export default function ProjectForm() {
   const {t} = useTranslation();
   const {values, handleChange} = useFormikContext<ProjectFormValues>();
 
   return (
-    <Column space={5}>
+    <Column>
       <FormTextField<ProjectFormValues>
         name="name"
         maxLength={PROJECT_NAME_MAX_LENGTH}
