+++ dev-client/src/translations/ka.json
@@ -332,7 +332,6 @@
         "form": {
             "name_min_length_error": "პროექტის სახელი უნდა შეიცავდეს მინიმუმ {{min}} სიმბოლოს",
             "name_max_length_error": "პროექტის სახელი უნდა შეიცავდეს მაქსიმუმ {{max}} სიმბოლოს",
-            "description_min_length_error": "პროექტის აღწერა უნდა იყოს მინიმუმ {{min}} სიმბოლო",
             "description_max_length_error": "პროექტის აღწერა უნდა შეიცავდეს მაქსიმუმ {{max}} სიმბოლოს"
         },
         "sort_label": "პროექტების დახარისხება",
