+++ dev-client/src/translations/uk.json
@@ -332,7 +332,6 @@
         "form": {
             "name_min_length_error": "Назва проєкту має містити щонайменше {{min}} символів",
             "name_max_length_error": "Назва проєкту має містити щонайбільше {{max}} символів",
-            "description_min_length_error": "Опис проекту має містити щонайменше {{min}} символів",
             "description_max_length_error": "Опис проекту має містити щонайбільше {{max}} символів"
         },
         "sort_label": "Сортувати проекти за",
